promote RDFGraph to a class from a typedef
also differentiate between usage of the RDFGraph class and of just a vector of RDFTriples in RDFSerializationProcessor.cpp. only the usage of an RDFGraph object needs to be able to filter out duplicate triple from storage.

also fix small bug in Literal constructor.

fixes toRDF test e036 because that test's result file has duplicate triples in it that we weren't ignoring before.