store pointers to nodeclusters in nodemapper; fixes toRDF tests tli09 and tli10
Before, when we were storing NodeCluster objects in the NodeMapper, we
might have issues when too many NodeClusters were added, causing the
vector we were using to increase capacity and invalidate all references
to those NodeClusters. Now we store pointers in a map, so we no longer
have reference invalidation problems.